Conflicts resolved in:
- Vagrantfile
- bin/common/install_certstrap.sh
- bin/dev/install_tool.sh

+++ bin/common/install_certstrap.sh
@@ -4,4 +4,6 @@
 echo "Installing certstrap ..."
 # We run chown in docker to avoid requiring sudo
 docker run --rm -v /usr/local/bin:/out:rw "golang:1.7" /usr/bin/env GOBIN=/out go get github.com/square/certstrap
-docker run --rm -v /usr/local/bin:/out:rw "golang:1.7" /bin/chown "$(id -u):$(id -g)" /out/certstrap
+if [[ $(stat -c '%u' "/usr/local/bin/certstrap") -eq 0 ]]; then
+  docker run --rm -v /usr/local/bin:/out:rw "golang:1.7" /bin/chown "$(id -u):$(id -g)" /out/certstrap
+fi
